pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ebay-rest"
version = "0.1.0"
description = "A modern Python SDK for eBay's REST APIs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Gustavo Belaunde", email = "gbelaunderojas@gmail.com"},
]
keywords = ["ebay", "sdk", "api", "rest", "ecommerce"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/GustavoBelaunde2004/ebay-rest"
Documentation = "https://github.com/GustavoBelaunde2004/ebay-rest#readme"
Repository = "https://github.com/GustavoBelaunde2004/ebay-rest"
Issues = "https://github.com/GustavoBelaunde2004/ebay-rest/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "requires_credentials: marks tests that require EBAY_CLIENT_ID and EBAY_CLIENT_SECRET",
    "requires_user_token: marks tests that require EBAY_USER_ACCESS_TOKEN",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["ebay_rest"]
omit = [
    "tests/*",
    "examples/*",
    "*/__pycache__/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

